--- app/src/main/java/com/lalilu/lmusic/apis/ApisModule.kt
@@ -1,30 +1,0 @@
-package com.lalilu.lmusic.apis
-
-import com.lalilu.lmusic.Config
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-class ApisModule {
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(Config.BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideNetWorkLyricService(retrofit: Retrofit): NeteaseDataSource {
-        return retrofit.create(NeteaseDataSource::class.java)
-    }
-}
