+++ app/src/main/java/com/lalilu/lmusic/service/MSongBrowser.kt
@@ -30,17 +30,6 @@ import javax.inject.Inject
 import javax.inject.Singleton
 import kotlin.coroutines.CoroutineContext
 
-interface EnhanceBrowser {
-    fun togglePlay(): Boolean
-    fun playByUri(uri: Uri): Boolean
-    fun playById(mediaId: String): Boolean
-    fun playById(mediaId: String, playWhenReady: Boolean): Boolean
-    fun addToNext(mediaId: String): Boolean
-    fun removeById(mediaId: String): Boolean
-    fun revokeRemove(): Boolean
-    fun moveByDelta(mediaId: String, delta: Int): Boolean
-}
-
 @Singleton
 class MSongBrowser @Inject constructor(
     @ApplicationContext
